MinIO blob tests

+++ implementations/dotnet/src/PolyPersist.Net.BlobStore.GridFS/GridFS_BlobContainer.cs
@@ -2,6 +2,7 @@
 using MongoDB.Driver;
 using MongoDB.Driver.GridFS;
 using PolyPersist.Net.Common;
+using System.Text;
 
 namespace PolyPersist.Net.BlobStore.GridFS
 {
@@ -27,15 +28,28 @@ namespace PolyPersist.Net.BlobStore.GridFS
         /// <inheritdoc/>
         async Task IBlobContainer<TBlob>.Upload(TBlob blob, Stream content)
         {
+            if (content == null || content.CanRead == false)
+                throw new Exception($"Blob '{typeof(TBlob).Name}' {blob.id} content cannot be read");
+
             await CollectionCommon.CheckBeforeInsert(blob).ConfigureAwait(false);
 
+            if (string.IsNullOrEmpty(blob.id) == true)
+                blob.id = Guid.NewGuid().ToString();
             blob.etag = Guid.NewGuid().ToString();
+            blob.LastUpdate = DateTime.UtcNow;
 
-            // insert metadata
-            await _metadataCollection.InsertOneAsync(blob).ConfigureAwait(false);
+            try
+            {
+                // insert metadata
+                await _metadataCollection.InsertOneAsync(blob).ConfigureAwait(false);
+            }
+            catch (MongoWriteException ex)
+            {
+                throw new Exception($"Blob '{typeof(TBlob).Name}' {blob.id} cannot be uploaded, beacuse of duplicate key", ex);
+            }
 
             // Upload empty content to GridFS, to create a entity is database
-            await _gridFSBucket.UploadFromStreamAsync(_makeId(blob), blob.fileName, content).ConfigureAwait(false);
+            await _gridFSBucket.UploadFromStreamAsync(_makeId(blob), content).ConfigureAwait(false);
         }
 
         /// <inheritdoc/>
@@ -56,7 +70,7 @@ namespace PolyPersist.Net.BlobStore.GridFS
                 throw new Exception($"Blob '{typeof(TBlob).Name}' {id} can not be removed. Database refused to acknowledge the operation.");
 
             if (result.DeletedCount != 1)
-                throw new Exception($"Blob '{typeof(TBlob).Name}'{id} can not be removed because it is already removed or changed.");
+                throw new Exception($"Blob '{typeof(TBlob).Name}' {id} can not be removed because it is does not exist");
 
             GridFSFileInfo fileInfo = await _getFileInfo(partitionKey, id).ConfigureAwait(false);
             if (fileInfo == null)
@@ -77,13 +91,16 @@ namespace PolyPersist.Net.BlobStore.GridFS
         /// <inheritdoc/>
         async Task IBlobContainer<TBlob>.UpdateContent(TBlob blob, Stream content)
         {
+            if (content == null || content.CanRead == false)
+                throw new Exception($"Blob '{typeof(TBlob).Name}' {blob.id} content cannot be read");
+
             GridFSFileInfo fileInfo = await _getFileInfo(blob.PartitionKey, blob.id).ConfigureAwait(false);
             if (fileInfo == null)
                 throw new Exception($"Blob '{typeof(TBlob).Name}' {blob.id} can not upload, because it is does not exist");
 
             await _gridFSBucket.DeleteAsync(fileInfo.Id).ConfigureAwait(false);
 
-            await _gridFSBucket.UploadFromStreamAsync(_makeId(blob), blob.fileName, content).ConfigureAwait(false);
+            await _gridFSBucket.UploadFromStreamAsync(_makeId(blob), content).ConfigureAwait(false);
         }
 
         /// <inheritdoc/>
@@ -102,7 +119,7 @@ namespace PolyPersist.Net.BlobStore.GridFS
             if (blob == null)
                 throw new Exception($"Entity '{typeof(TBlob).Name}' {blob.id} can not be updated because it is already changed");
         }
-    
+
 
         /// <inheritdoc/>
         object IBlobContainer<TBlob>.GetUnderlyingImplementation()
@@ -110,15 +127,25 @@ namespace PolyPersist.Net.BlobStore.GridFS
             return _gridFSBucket;
         }
 
-        private ObjectId _makeId(TBlob blob)
+        private string _makeId(TBlob blob)
             => _makeId(blob.PartitionKey, blob.id);
 
-        private ObjectId _makeId(string partitionKey, string id)
-            => new ObjectId($"{partitionKey}-{id}");
+        private string _makeId(string partitionKey, string id)
+            => $"{partitionKey}-{id}";
 
+        static string StringToHex(string input)
+        {
+            byte[] bytes = Encoding.UTF8.GetBytes(input);
+            StringBuilder hex = new StringBuilder(bytes.Length * 2);
+            foreach (byte b in bytes)
+            {
+                hex.AppendFormat("{0:x2}", b); // kisbetűs hex; használd {0:X2} ha nagybetűs kell
+            }
+            return hex.ToString();
+        }
         private async Task<GridFSFileInfo> _getFileInfo(string partitionKey, string id)
         {
-            IAsyncCursor<GridFSFileInfo> cursor = await _filesCollection.FindAsync(fi => fi.Id == _makeId(partitionKey, id)).ConfigureAwait(false);
+            IAsyncCursor<GridFSFileInfo> cursor = await _filesCollection.FindAsync(fi => fi.Filename == _makeId(partitionKey, id)).ConfigureAwait(false);
             return await cursor.FirstOrDefaultAsync().ConfigureAwait(false);
         }
     }
